+++ examples/viewer/config/webpack.dev.js
@@ -9,6 +9,7 @@ const devConfig = {
   },
   serve: {
     port: 3001,
+    host: '0.0.0.0',
     hot: true,
     clipboard: false,
     dev: {